% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSRT.R
\name{intSSRT}
\alias{intSSRT}
\title{intSSRT}
\usage{
intSSRT(dat_tmp)
}
\arguments{
\item{dat_tmp}{a dataframe containing vp number, rt, accuracy, the condition
and stop-signal delays}

\item{vp_num}{Name of the column containing the vp identifier.}

\item{cond}{Name of the column that reflects the current condition stop =
c(3, 4); go = c(1, 2)}

\item{correct_code}{Name of the column containing the accuracy (1 = correct)}

\item{ssd}{Name of the column containing the current stop signal delay}
}
\value{
a dataframe containing an estimate of the SSRT, the mean reaction
times on go trials, the mean stop signal delay and the mean p(reponse |
signal) for each VP
}
\description{
utilises the RT distribution of correct Go Trials in a
Stop-Signal task to estimate the Stop-Signal Reaction Time
}
\examples{
dat <- meanSSRT(dat)
}
