% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSRT.R
\name{meanSSRT}
\alias{meanSSRT}
\title{meanSSRT}
\usage{
meanSSRT(dat_tmp)
}
\arguments{
\item{dat_tmp}{a dataframe containing vp number, rt, accuracy, the condition
and stop-signal delays}

\item{vp_num}{Name of the column containing the vp identifier.}

\item{cond}{Name of the column that reflects the current condition stop =
c(3, 4); go = c(1, 2)}

\item{correct_code}{Name of the column containing the accuracy (1 = correct)}

\item{ssd}{Name of the column containing the current stop signal delay}
}
\value{
a dataframe containing an estimate of the SSRT, the mean reaction
times on go trials and the mean stop signal delay for each VP
}
\description{
Estimates the SSRT using the mean of all SSDs in the
experiment. This calculation hinges upon the assumption that p(response |
signal = 0.5)
}
\examples{
dat <- meanSSRT(dat)
}
